requirements.txt
anthropic
elevenlabs
requests
moviepy
openai-whisper
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
fal-client
